- moved the desi_simqso_model to py/deisim/scripts until a better place is defined - added an option to use the modified simqso_model to  _make_simqso_templates and pass that option from quickquasars.
